fix(runtime): split DataView registry module
